requirements.txt
google-genai>=1.24.0
pandas>=2.3.0
pdfplumber>=0.11.7
pymupdf>=1.26.1
python-docx>=1.2.0
streamlit>=1.46.1 